+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.8.2`
+
+- **added** Makie extensions for bivariate and VSUP figures [#124]
+- **added** `clamp` and `clamp!` for layers
+
 ## `v1.8.1` 
 
 - **added** a `MakieExtension` for graticules